Network.AWS.SMS.GetReplicationRuns

Description

The GetReplicationRuns API will return all ReplicationRuns for a given ReplicationJob. This API returns a paginated list, that may be consecutively called with nextToken to retrieve all ReplicationRuns for a ReplicationJob.

This operation returns paginated results.

Creating a Request

getReplicationRuns

Arguments

:: Text
grrReplicationJobId

-> GetReplicationRuns

Creates a value of GetReplicationRuns with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grrNextToken - Undocumented member.

grrMaxResults - Undocumented member.

grrReplicationJobId - Undocumented member.

data GetReplicationRuns

See: getReplicationRuns smart constructor.

Request Lenses

grrNextToken :: Lens' GetReplicationRuns (Maybe Text)

Undocumented member.

grrMaxResults :: Lens' GetReplicationRuns (Maybe Int)

Undocumented member.

grrReplicationJobId :: Lens' GetReplicationRuns Text

Undocumented member.

Destructuring the Response

getReplicationRunsResponse

Arguments

:: Int
grrrsResponseStatus

-> GetReplicationRunsResponse

Creates a value of GetReplicationRunsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grrrsReplicationJob - Undocumented member.

grrrsNextToken - Undocumented member.

grrrsReplicationRunList - Undocumented member.

grrrsResponseStatus - -- | The response status code.

data GetReplicationRunsResponse

See: getReplicationRunsResponse smart constructor.

Response Lenses

grrrsReplicationJob :: Lens' GetReplicationRunsResponse (Maybe ReplicationJob)

Undocumented member.

grrrsNextToken :: Lens' GetReplicationRunsResponse (Maybe Text)

Undocumented member.

grrrsReplicationRunList :: Lens' GetReplicationRunsResponse [ReplicationRun]

Undocumented member.

grrrsResponseStatus :: Lens' GetReplicationRunsResponse Int

- | The response status code.