Network.AWS.SMS.GetServers

Description

The GetServers API returns a list of all servers in your server catalog. For this call to succeed, you must previously have called ImportServerCatalog.

This operation returns paginated results.

Creating a Request

getServers :: GetServers

Creates a value of GetServers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsNextToken - Undocumented member.

gsMaxResults - Undocumented member.

data GetServers

See: getServers smart constructor.

Request Lenses

gsNextToken :: Lens' GetServers (Maybe Text)

Undocumented member.

gsMaxResults :: Lens' GetServers (Maybe Int)

Undocumented member.

Destructuring the Response

getServersResponse

Arguments

:: Int
gsrsResponseStatus

-> GetServersResponse

Creates a value of GetServersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsrsServerCatalogStatus - Undocumented member.

gsrsLastModifiedOn - Undocumented member.

gsrsNextToken - Undocumented member.

gsrsServerList - Undocumented member.

gsrsResponseStatus - -- | The response status code.

data GetServersResponse

See: getServersResponse smart constructor.

Response Lenses

gsrsServerCatalogStatus :: Lens' GetServersResponse (Maybe ServerCatalogStatus)

Undocumented member.

gsrsLastModifiedOn :: Lens' GetServersResponse (Maybe UTCTime)

Undocumented member.

gsrsNextToken :: Lens' GetServersResponse (Maybe Text)

Undocumented member.

gsrsServerList :: Lens' GetServersResponse [Server]

Undocumented member.

gsrsResponseStatus :: Lens' GetServersResponse Int

- | The response status code.