Network.AWS.SMS.ImportServerCatalog

Description

The ImportServerCatalog API is used to gather the complete list of on-premises servers on your premises. This API call requires connectors to be installed and monitoring all servers you would like imported. This API call returns immediately, but may take some time to retrieve all of the servers.

Creating a Request

importServerCatalog :: ImportServerCatalog

Creates a value of ImportServerCatalog with the minimum fields required to make a request.

data ImportServerCatalog

See: importServerCatalog smart constructor.

Destructuring the Response

importServerCatalogResponse

Arguments

:: Int
iscrsResponseStatus

-> ImportServerCatalogResponse

Creates a value of ImportServerCatalogResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iscrsResponseStatus - -- | The response status code.

data ImportServerCatalogResponse

See: importServerCatalogResponse smart constructor.

Response Lenses

iscrsResponseStatus :: Lens' ImportServerCatalogResponse Int

- | The response status code.