Network.AWS.SMS.Types

Description

Service Configuration

sms :: Service

API version 2016-10-24 of the Amazon Server Migration Service SDK configuration.

Errors

_ReplicationRunLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

This user has exceeded the maximum allowed Replication Run limit.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter specified in the request is not valid, is unsupported, or cannot be used.

_NoConnectorsAvailableException :: AsError a => Getting (First ServiceError) a ServiceError

No connectors are available to handle this request. Please associate connector(s) and verify any existing connectors are healthy and can respond to requests.

_ReplicationJobNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified Replication Job cannot be found.

_ServerCannotBeReplicatedException :: AsError a => Getting (First ServiceError) a ServiceError

The provided server cannot be replicated.

_InternalError :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occured.

_ReplicationJobAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

An active Replication Job already exists for the specified server.

_OperationNotPermittedException :: AsError a => Getting (First ServiceError) a ServiceError

The specified operation is not allowed. This error can occur for a number of reasons; for example, you might be trying to start a Replication Run before seed Replication Run.

_MissingRequiredParameterException :: AsError a => Getting (First ServiceError) a ServiceError

The request is missing a required parameter. Ensure that you have supplied all the required parameters for the request.

_UnauthorizedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

This user does not have permissions to perform this operation.

ConnectorCapability

data ConnectorCapability

Capabilities for a Connector

Constructors

CCVsphere

ConnectorStatus

data ConnectorStatus

Status of on-premise Connector

Constructors

Healthy

Unhealthy

LicenseType

data LicenseType

The license type to be used for the Amazon Machine Image (AMI) created after a successful ReplicationRun.

Constructors

AWS

Byol

ReplicationJobState

data ReplicationJobState

Current state of Replication Job

Constructors

RJSActive

RJSDeleted

RJSDeleting

RJSFailed

RJSPending

ReplicationRunState

data ReplicationRunState

Current state of Replication Run

Constructors

RRSActive

RRSCompleted

RRSDeleted

RRSDeleting

RRSFailed

RRSMissed

RRSPending

ReplicationRunType

data ReplicationRunType

Type of Replication Run

Constructors

Automatic

OnDemand

ServerCatalogStatus

data ServerCatalogStatus

Status of Server catalog

Constructors

Available

Deleted

Expired

Importing

NotImported

ServerType

data ServerType

Type of server.

Constructors

VirtualMachine

VMManagerType

data VMManagerType

VM Management Product

Constructors

Vsphere

Connector

data Connector

Object representing a Connector

See: connector smart constructor.

connector :: Connector

Creates a value of Connector with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cStatus - Undocumented member.

cVmManagerName - Undocumented member.

cIpAddress - Undocumented member.

cVmManagerId - Undocumented member.

cVmManagerType - Undocumented member.

cConnectorId - Undocumented member.

cAssociatedOn - Undocumented member.

cMacAddress - Undocumented member.

cVersion - Undocumented member.

cCapabilityList - Undocumented member.

cStatus :: Lens' Connector (Maybe ConnectorStatus)

Undocumented member.

cVmManagerName :: Lens' Connector (Maybe Text)

Undocumented member.

cIpAddress :: Lens' Connector (Maybe Text)

Undocumented member.

cVmManagerId :: Lens' Connector (Maybe Text)

Undocumented member.

cVmManagerType :: Lens' Connector (Maybe VMManagerType)

Undocumented member.

cConnectorId :: Lens' Connector (Maybe Text)

Undocumented member.

cAssociatedOn :: Lens' Connector (Maybe UTCTime)

Undocumented member.

cMacAddress :: Lens' Connector (Maybe Text)

Undocumented member.

cVersion :: Lens' Connector (Maybe Text)

Undocumented member.

cCapabilityList :: Lens' Connector [ConnectorCapability]

Undocumented member.

ReplicationJob

data ReplicationJob

Object representing a Replication Job

See: replicationJob smart constructor.

replicationJob :: ReplicationJob

Creates a value of ReplicationJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rjFrequency - Undocumented member.

rjState - Undocumented member.

rjServerType - Undocumented member.

rjServerId - Undocumented member.

rjLicenseType - Undocumented member.

rjRoleName - Undocumented member.

rjVmServer - Undocumented member.

rjReplicationJobId - Undocumented member.

rjReplicationRunList - Undocumented member.

rjNextReplicationRunStartTime - Undocumented member.

rjStatusMessage - Undocumented member.

rjLatestAMIId - Undocumented member.

rjSeedReplicationTime - Undocumented member.

rjDescription - Undocumented member.

rjFrequency :: Lens' ReplicationJob (Maybe Int)

Undocumented member.

rjState :: Lens' ReplicationJob (Maybe ReplicationJobState)

Undocumented member.

rjServerType :: Lens' ReplicationJob (Maybe ServerType)

Undocumented member.

rjServerId :: Lens' ReplicationJob (Maybe Text)

Undocumented member.

rjLicenseType :: Lens' ReplicationJob (Maybe LicenseType)

Undocumented member.

rjRoleName :: Lens' ReplicationJob (Maybe Text)

Undocumented member.

rjVmServer :: Lens' ReplicationJob (Maybe VMServer)

Undocumented member.

rjReplicationJobId :: Lens' ReplicationJob (Maybe Text)

Undocumented member.

rjReplicationRunList :: Lens' ReplicationJob [ReplicationRun]

Undocumented member.

rjNextReplicationRunStartTime :: Lens' ReplicationJob (Maybe UTCTime)

Undocumented member.

rjStatusMessage :: Lens' ReplicationJob (Maybe Text)

Undocumented member.

rjLatestAMIId :: Lens' ReplicationJob (Maybe Text)

Undocumented member.

rjSeedReplicationTime :: Lens' ReplicationJob (Maybe UTCTime)

Undocumented member.

rjDescription :: Lens' ReplicationJob (Maybe Text)

Undocumented member.

ReplicationRun

data ReplicationRun

Object representing a Replication Run

See: replicationRun smart constructor.

replicationRun :: ReplicationRun

Creates a value of ReplicationRun with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrState - Undocumented member.

rrReplicationRunId - Undocumented member.

rrScheduledStartTime - Undocumented member.

rrStatusMessage - Undocumented member.

rrCompletedTime - Undocumented member.

rrAmiId - Undocumented member.

rrType - Undocumented member.

rrDescription - Undocumented member.

rrState :: Lens' ReplicationRun (Maybe ReplicationRunState)

Undocumented member.

rrReplicationRunId :: Lens' ReplicationRun (Maybe Text)

Undocumented member.

rrScheduledStartTime :: Lens' ReplicationRun (Maybe UTCTime)

Undocumented member.

rrStatusMessage :: Lens' ReplicationRun (Maybe Text)

Undocumented member.

rrCompletedTime :: Lens' ReplicationRun (Maybe UTCTime)

Undocumented member.

rrAmiId :: Lens' ReplicationRun (Maybe Text)

Undocumented member.

rrType :: Lens' ReplicationRun (Maybe ReplicationRunType)

Undocumented member.

rrDescription :: Lens' ReplicationRun (Maybe Text)

Undocumented member.

Server

data Server

Object representing a server

See: server smart constructor.

server :: Server

Creates a value of Server with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sServerType - Undocumented member.

sServerId - Undocumented member.

sReplicationJobTerminated - Undocumented member.

sVmServer - Undocumented member.

sReplicationJobId - Undocumented member.

sServerType :: Lens' Server (Maybe ServerType)

Undocumented member.

sServerId :: Lens' Server (Maybe Text)

Undocumented member.

sReplicationJobTerminated :: Lens' Server (Maybe Bool)

Undocumented member.

sVmServer :: Lens' Server (Maybe VMServer)

Undocumented member.

sReplicationJobId :: Lens' Server (Maybe Text)

Undocumented member.

VMServer

data VMServer

Object representing a VM server

See: vMServer smart constructor.

vMServer :: VMServer

Creates a value of VMServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vmsVmManagerName - Undocumented member.

vmsVmManagerType - Undocumented member.

vmsVmServerAddress - Undocumented member.

vmsVmName - Undocumented member.

vmsVmPath - Undocumented member.

vmsVmManagerName :: Lens' VMServer (Maybe Text)

Undocumented member.

vmsVmManagerType :: Lens' VMServer (Maybe VMManagerType)

Undocumented member.

vmsVmServerAddress :: Lens' VMServer (Maybe VMServerAddress)

Undocumented member.

vmsVmName :: Lens' VMServer (Maybe Text)

Undocumented member.

vmsVmPath :: Lens' VMServer (Maybe Text)

Undocumented member.

VMServerAddress

data VMServerAddress

Object representing a server's location

See: vMServerAddress smart constructor.

vMServerAddress :: VMServerAddress

Creates a value of VMServerAddress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vmsaVmManagerId - Undocumented member.

vmsaVmId - Undocumented member.

vmsaVmManagerId :: Lens' VMServerAddress (Maybe Text)

Undocumented member.

vmsaVmId :: Lens' VMServerAddress (Maybe Text)

Undocumented member.