Network.AWS.SMS.UpdateReplicationJob

Description

The UpdateReplicationJob API is used to change the settings of your existing ReplicationJob created using CreateReplicationJob. Calling this API will affect the next scheduled ReplicationRun.

Creating a Request

updateReplicationJob

Arguments

:: Text
urjReplicationJobId

-> UpdateReplicationJob

Creates a value of UpdateReplicationJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urjFrequency - Undocumented member.

urjLicenseType - Undocumented member.

urjRoleName - Undocumented member.

urjNextReplicationRunStartTime - Undocumented member.

urjDescription - Undocumented member.

urjReplicationJobId - Undocumented member.

data UpdateReplicationJob

See: updateReplicationJob smart constructor.

Request Lenses

urjFrequency :: Lens' UpdateReplicationJob (Maybe Int)

Undocumented member.

urjLicenseType :: Lens' UpdateReplicationJob (Maybe LicenseType)

Undocumented member.

urjRoleName :: Lens' UpdateReplicationJob (Maybe Text)

Undocumented member.

urjNextReplicationRunStartTime :: Lens' UpdateReplicationJob (Maybe UTCTime)

Undocumented member.

urjDescription :: Lens' UpdateReplicationJob (Maybe Text)

Undocumented member.

urjReplicationJobId :: Lens' UpdateReplicationJob Text

Undocumented member.

Destructuring the Response

updateReplicationJobResponse

Arguments

:: Int
urjrsResponseStatus

-> UpdateReplicationJobResponse

Creates a value of UpdateReplicationJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urjrsResponseStatus - -- | The response status code.

data UpdateReplicationJobResponse

See: updateReplicationJobResponse smart constructor.

Response Lenses

urjrsResponseStatus :: Lens' UpdateReplicationJobResponse Int

- | The response status code.